rewrote rendering and allow GO to have a float pos

+++ vroom/game_object.py
@@ -7,7 +7,7 @@ class GameObject:
         self.name: str = name
         self.components: list[Component] = []
         self.id = uuid.uuid4()
-        self.pos: tuple[int, int] = 0, 0
+        self.pos: tuple[float, float] = 0, 0
         self.rotation: float = 0
         self.scale: float = 1
